@@ -1,5 +1,5 @@
 [options]
-python_requires = >= 3.7
+python_requires = >= 3.11
 
 [flake8]
 max-line-length = 120
@@ -13,7 +13,7 @@ ignore = W503, E203, E231, E501
 jobs = 4
 
 [mypy]
-python_version = 3.7
+python_version = 3.11
 ignore_missing_imports = True
 check_untyped_defs = True
 disallow_untyped_defs = True
