@@ -1,6 +1,5 @@
 import logging
 import os
-from typing import List
 
 from wiremind_kubernetes import KubernetesDeploymentManager
 
@@ -18,7 +17,7 @@ def ensure_safe_run() -> None:
     """
     package_v: str = get_version()
     # Get "1.2" from "1.2.3"
-    package_v_major_minor: List[str] = package_v.split(".", 2)[:2]
+    package_v_major_minor: list[str] = package_v.split(".", 2)[:2]
 
     helm_chart_v: str = os.getenv("HELM_CHART_VERSION", "")
     if not helm_chart_v:
@@ -26,7 +25,7 @@ def ensure_safe_run() -> None:
             "Couldn't get the Chartreuse's Helm Chart version from the env var HELM_CHART_VERSION,"
             " couldn't make sure that the package is of a compatible version, ABORTING!"
         )
-    helm_chart_v_major_minor: List[str] = helm_chart_v.split(".", 2)[:2]
+    helm_chart_v_major_minor: list[str] = helm_chart_v.split(".", 2)[:2]
     if helm_chart_v_major_minor != package_v_major_minor:
         raise ValueError(
             f"Chartreuse's Helm Chart version '{helm_chart_v}' and the package's version '{package_v}' "
