@@ -2,7 +2,7 @@ import logging
 import os
 import subprocess
 import time
-from typing import Generator
+from collections.abc import Generator
 
 import pytest
 import sqlalchemy
