@@ -57,7 +57,7 @@ def run_migrations_online() -> None:
     """
     patroni_postgresql: bool = "patroni_postgresql" in context.get_x_argument(as_dictionary=True)
     connectable = engine_from_config(
-        config.get_section(config.config_ini_section),
+        config.get_section(config.config_ini_section),  # type: ignore
         prefix="sqlalchemy.",
         poolclass=pool.NullPool,
     )
